api/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
dj-rest-auth==4.0.1
Django==4.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.5
Faker==18.7.0
filelock==3.12.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.2
mypy-extensions==0.4.3
packaging==23.1
pathspec==0.9.0
platformdirs==2.4.1
psycopg==3.1.9
psycopg-binary==3.1.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0
zipp==3.8.0
